delete LoggedInScreen; rename HomeScreen to Auth and ButtonScreen to Home
