+++ frontend/App.js
@@ -1,9 +1,8 @@
 import React, { Component }  from 'react';
 import { StyleSheet, Text, View, Button } from 'react-native';
 import { StackNavigator } from 'react-navigation';
+import AuthScreen from './src/screens/AuthScreen';
 import HomeScreen from './src/screens/HomeScreen';
-import ButtonScreen from './src/screens/ButtonScreen';
-import LoggedInScreen from './src/screens/LoggedInScreen';
 import ListScreen from './src/screens/ListScreen';
 
 
@@ -16,8 +15,7 @@ export default class App extends Component {
 }
 
 const AppNavigator = new StackNavigator({
+  AuthScreen: { screen: AuthScreen },
   HomeScreen: { screen: HomeScreen },
-  LoggedInScreen: { screen: LoggedInScreen },
-  ButtonScreen: { screen: ButtonScreen },
   ListScreen: { screen: ListScreen }
 });
