+++ frontend/src/__tests__/screens/AuthScreen.test.js
@@ -1,24 +1,24 @@
 import React from "react";
 import {shallow} from 'enzyme';
-import LoggedInScreen from "../../screens/LoggedInScreen";
+import AuthScreen from "../../screens/AuthScreen";
 import renderer from "react-test-renderer";
 
-describe ("LoggedInScreen", () => {
+describe ("AuthScreen", () => {
   it("renders without crashing", () => {
-    const rendered = renderer.create(<LoggedInScreen />).toJSON();
+    const rendered = renderer.create(<AuthScreen />).toJSON();
     expect(rendered).toBeTruthy();
     expect(rendered).toMatchSnapshot();
   });
 
-  it('calls navigate on pressing go to button', () => {
+  it('calls navigate on pressing sign in', () => {
     const fakeNavigation = {navigate: jest.fn()};
 
-    let loggedIn = renderer.create(<LoggedInScreen
+    let authScreen = renderer.create(<AuthScreen
       navigation={fakeNavigation}
       />
     ).getInstance();
 
-    loggedIn.handleButtonScreenPress();
+    authScreen.handleLogInPress();
     expect(fakeNavigation.navigate).toHaveBeenCalledTimes(1);
   })
 })
